Added check if .env.test exists

+++ setup.js
@@ -1,7 +1,9 @@
 fs = require('fs');
-let stream = fs.createReadStream('.env.test')
-  .pipe(fs.createWriteStream('.env'));
+if (fs.existsSync('.env.test')) {
+  let stream = fs.createReadStream('.env.test')
+    .pipe(fs.createWriteStream('.env'));
 
-stream.on('close', () => {
-  fs.unlink('.env.test');
-});
+  stream.on('close', () => {
+    fs.unlink('.env.test');
+  });    
+}
